Small Fixes for PEP8 and Fixes for render_settings path

+++ server/modules/shots.py
@@ -3,7 +3,6 @@ from flask import Blueprint, render_template, abort, jsonify, request
 from model import *
 from jobs import *
 from utils import *
-from decimal import Decimal
 
 shots_module = Blueprint('shots_module', __name__)
 
@@ -20,13 +19,12 @@ def delete_shot(shot_id):
 
 @shots_module.route('/shots/')
 def shots():
-    from decimal import Decimal
     shots = {}
     for shot in Shots.select():
         percentage_done = 0
         frame_count = shot.frame_end - shot.frame_start + 1
         current_frame = shot.current_frame - shot.frame_start + 1
-        percentage_done = Decimal(current_frame) / Decimal(frame_count) * Decimal(100)
+        percentage_done = float(current_frame) / float(frame_count) * float(100)
         percentage_done = round(percentage_done, 1)
 
         if percentage_done == 100:
@@ -97,7 +95,6 @@ def shot_reset(shot_id):
     except Shots.DoesNotExist:
         shot = None
         print('[error] Shot not found')
-        
         return 'Shot %d not found' % shot_id
 
     if shot.status == 'running':
@@ -112,7 +109,6 @@ def shot_reset(shot_id):
         return jsonify(
             shot_id=shot.id,
             status='ready')
-
 
 
 @shots_module.route('/shots/add', methods=['POST'])
